+++ src/buf/bits.rs
@@ -1,23 +1,15 @@
 //! Provides a single-bit, 2D grid data structure backed by a linear buffer.
 //!
 //! The main type, [`GridBits`], is highly generic, allowing it to act as a view over any buffer
-//! that can be treated as a slice.
-//!
-//! ## Convenience Types
-//!
-//! For ease of use, several type aliases are provided for common use cases:
-//! - [`VecBits`]: An owned grid backed by a `Vec<u8>`
-//! - [`ArrayBits`]: An owned grid backed by a fixed-size array, `[u8; N]`.
-//! - [`SliceBits`]: A read-only, borrowed view over an existing slice.
-//! - [`SliceMutBits`]: A mutable, borrowed view over an existing slice.
+//! that can be treated as a slice of bytes.
 //!
 //! # Examples
 //!
-//! Creating an owned `VecGridBits` and accessing an element:
+//! Creating an owned `GridBits` and accessing an element:
 //! ```
-//! use grixy::{core::{Pos, RowMajor}, buf::bits::VecBits};
+//! use grixy::{core::{Pos, RowMajor}, buf::bits::GridBits};
 //!
-//! let grid = VecBits::<u8, RowMajor>::new(8, 1);
+//! let grid = GridBits::<u8, _, RowMajor>::new(8, 1);
 //! assert_eq!(grid.get(Pos::new(3, 0)), Some(false));
 //! ```
 
@@ -39,7 +31,7 @@ extern crate alloc;
 ///
 /// The grid is stored in a linear buffer, with elements accessed in an order defined by [`Layout`].
 #[derive(Debug, Clone)]
-pub struct GridBits<T, B, L>
+pub struct GridBits<T, B, L = RowMajor>
 where
     T: BitOps,
     L: Layout,
@@ -51,37 +43,90 @@ where
     _layout: PhantomData<L>,
 }
 
-impl<T, B> GridBits<T, B, RowMajor>
+impl<T, B, L> GridBits<T, B, L>
 where
     T: BitOps,
     B: AsRef<[T]>,
+    L: Layout,
 {
-    /// Creates a `GridBuf` using an existing data buffer, specifying the grid dimensions.
+    /// Returns a grid from an existing buffer with a given width in columns.
     ///
-    /// The maximum width that can be used is determined by [`BitOps::MAX_WIDTH`].
+    /// The height is inferred from the buffer length and width.
     ///
-    /// The data buffer is expected to be in [`RowMajor`] order.
+    /// Any data type that can be represented as a slice can be used as the buffer type, including
+    /// arrays, slices, and vectors.
     ///
-    /// ## Errors
+    /// ## Panics
     ///
-    /// Returns an error if the buffer size does not match the expected size.
-    pub fn with_buffer_row_major(
-        buffer: B,
-        width: usize,
-        height: usize,
-    ) -> Result<Self, GridError> {
-        Self::with_buffer(buffer, width, height)
+    /// This panics if the buffer length is not a multiple of the width.
+    ///
+    /// ## Example
+    ///
+    /// ```rust
+    /// use grixy::{core::Pos, buf::bits::GridBits};
+    ///
+    /// let grid = GridBits::<_, Vec<u8>>::from_buffer(vec![1, 2, 3, 4], 2);
+    /// assert_eq!(grid.get(Pos::new(0, 0)), Some(true));
+    ///
+    /// assert_eq!(grid.get(Pos::new(1, 0)), Some(false));
+    /// assert_eq!(grid.get(Pos::new(2, 0)), None);
+    /// ```
+    #[must_use]
+    pub fn from_buffer(buffer: B, width: usize) -> Self {
+        let height = buffer.as_ref().len() * T::MAX_WIDTH / width;
+        assert!(
+            height * width == buffer.as_ref().len() * T::MAX_WIDTH,
+            "Buffer length must be a multiple of width"
+        );
+        Self {
+            buffer,
+            width,
+            height,
+            _element: PhantomData,
+            _layout: PhantomData,
+        }
     }
+}
 
-    /// Creates a new `GridBuf` using an existing data buffer, specifying the grid dimensions.
+#[cfg(feature = "alloc")]
+impl<T> GridBits<T, alloc::vec::Vec<T>, RowMajor>
+where
+    T: BitOps + Default,
+{
+    /// Creates a new grid with the specified width and height, filled with a default value.
     ///
-    /// The data buffer is expected to be in [`RowMajor`] order.
+    /// This creates a grid with a row-major layout; see [`new_with_layout`] to customize.
     ///
-    /// ## Safety
+    /// ## Example
+    #[must_use]
+    pub fn new(width: usize, height: usize) -> Self {
+        let buffer = alloc::vec![T::default(); (width * height).div_ceil(T::MAX_WIDTH)];
+        Self::from_buffer(buffer, width)
+    }
+}
+
+#[cfg(feature = "alloc")]
+impl<T, L> GridBits<T, alloc::vec::Vec<T>, L>
+where
+    T: BitOps,
+    L: Layout,
+{
+    /// Creates a new grid with the specified width, height, and layout, filled with the default value.
     ///
-    /// The caller must ensure that the buffer is large enough to hold `width * height` elements.
-    pub unsafe fn with_buffer_row_major_unchecked(buffer: B, width: usize, height: usize) -> Self {
-        unsafe { Self::with_buffer_unchecked(buffer, width, height) }
+    /// # Example
+    /// ```
+    /// use grixy::{core::{Pos, RowMajor}, buf::bits::GridBits};
+    ///
+    /// let grid = GridBits::<u8, _, RowMajor>::new_with_layout(8, 1);
+    /// assert_eq!(grid.get(Pos::new(0, 0)), Some(false));
+    /// ```
+    #[must_use]
+    pub fn new_with_layout(width: usize, height: usize) -> Self
+    where
+        T: Default,
+    {
+        let buffer = alloc::vec![T::default(); (width * height).div_ceil(T::MAX_WIDTH)];
+        Self::from_buffer(buffer, width)
     }
 }
 
@@ -180,55 +225,6 @@ where
     }
 }
 
-/// A 2-dimensional grid implemented by a fixed-size array buffer of bytes.
-///
-/// This is a convenience type for using arrays as the underlying buffer.
-///
-/// ## Layout
-///
-/// The grid is stored in a linear buffer, with elements accessed in an order defined by [`Layout`].
-pub type ArrayBits<T, const N: usize, L> = GridBits<T, [T; N], L>;
-impl<T, const N: usize, L> GridBits<T, [T; N], L>
-where
-    T: BitOps,
-    L: Layout,
-{
-    /// Creates a new `GridBits` backed by a fixed-size array with the specified width and height.
-    ///
-    /// Each element is initialized to `false`.
-    ///
-    /// ## Panics
-    ///
-    /// Panics if the buffer size does not match the expected size.
-    #[must_use]
-    pub fn new(width: usize, height: usize) -> Self {
-        Self::new_filled(width, height, false)
-    }
-
-    /// Creates a new `GridBits` backed by a fixed-size array with the specified width and height.
-    ///
-    /// Each element is initialized to the provided value.
-    ///
-    /// ## Panics
-    ///
-    /// Panics if the buffer size does not match the expected size.
-    #[must_use]
-    pub fn new_filled(width: usize, height: usize, value: bool) -> Self {
-        let size = width * height;
-        assert!(
-            size.div_ceil(T::MAX_WIDTH) <= N,
-            "Buffer size does not match grid dimensions"
-        );
-        let mut buffer = [T::from_usize(0); N];
-        for i in 0..size {
-            if value {
-                buffer[i / T::MAX_WIDTH] |= T::from_usize(1 << (i % T::MAX_WIDTH));
-            }
-        }
-        unsafe { Self::with_buffer_unchecked(buffer, width, height) }
-    }
-}
-
 impl<T, B, L> GridBits<T, B, L>
 where
     T: BitOps,
@@ -257,64 +253,6 @@ where
     }
 }
 
-/// A 2-dimensional grid implemented by a slice buffer of bytes.
-///
-/// This is a convenience type for using slices as the underlying buffer.
-///
-/// ## Layout
-///
-/// The grid is stored in a linear buffer, with elements accessed in an order defined by [`Layout`].
-pub type SliceBits<'a, T, L> = GridBits<T, &'a [T], L>;
-
-/// A 2-dimensional grid implemented by a mutable slice buffer of bytes.
-///
-/// This is a convenience type for using mutable slices as the underlying buffer.
-///
-/// ## Layout
-///
-/// The grid is stored in a linear buffer, with elements accessed in an order defined by [`Layout`].
-pub type SliceMutBits<'a, T, L> = GridBits<T, &'a mut [T], L>;
-
-/// A 2-dimensional grid implemented by a vector buffer of bytes.
-///
-/// This is a convenience type for using `Vec` as the underlying buffer.
-///
-/// ## Layout
-///
-/// The grid is stored in a linear buffer, with elements accessed in an order defined by [`Layout`].
-#[cfg(feature = "alloc")]
-pub type VecBits<T, L> = GridBits<T, alloc::vec::Vec<T>, L>;
-
-#[cfg(feature = "alloc")]
-impl<T, L> GridBits<T, alloc::vec::Vec<T>, L>
-where
-    T: BitOps,
-    L: Layout,
-{
-    /// Creates a new `GridBits` backed by a `Vec` with the specified width and height.
-    ///
-    /// Each element is initialized to `false`.
-    #[must_use]
-    pub fn new(width: usize, height: usize) -> Self {
-        Self::new_filled(width, height, false)
-    }
-
-    /// Creates a new `GridBits` backed by a `Vec` with the specified width and height.
-    ///
-    /// Each element is initialized to the provided value.
-    #[must_use]
-    pub fn new_filled(width: usize, height: usize, value: bool) -> Self {
-        let size = width * height;
-        let mut buffer = alloc::vec![T::from_usize(0); size.div_ceil(T::MAX_WIDTH)];
-        for i in 0..size {
-            if value {
-                buffer[i / T::MAX_WIDTH] |= T::from_usize(1 << (i % T::MAX_WIDTH));
-            }
-        }
-        unsafe { Self::with_buffer_unchecked(buffer, width, height) }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     extern crate alloc;
@@ -322,14 +260,14 @@ mod tests {
     use ixy::index::RowMajor;
 
     use crate::{
-        buf::bits::{ArrayBits, SliceBits, SliceMutBits, VecBits},
+        buf::bits::GridBits,
         core::{GridError, Pos},
     };
 
     #[test]
     fn impl_arr() {
         let data: [u8; 1] = [0b0000_0001];
-        let grid = ArrayBits::<_, 1, RowMajor>::with_buffer(data, 8, 1).unwrap();
+        let grid = GridBits::<_, _, RowMajor>::from_buffer(data, 8);
 
         assert_eq!(grid.get(Pos::new(0, 0)), Some(true));
         assert_eq!(grid.get(Pos::new(1, 0)), Some(false));
@@ -339,7 +277,7 @@ mod tests {
 
     #[test]
     fn arr_new() {
-        let grid = ArrayBits::<u8, 1, RowMajor>::new(8, 1);
+        let grid = GridBits::<u8, _, RowMajor>::new(8, 1);
         assert_eq!(grid.get(Pos::new(0, 0)), Some(false));
         assert_eq!(grid.get(Pos::new(7, 0)), Some(false));
         assert_eq!(grid.get(Pos::new(8, 0)), None);
@@ -347,24 +285,15 @@ mod tests {
     }
 
     #[test]
-    fn arr_new_filled() {
-        let grid = ArrayBits::<u8, 1, RowMajor>::new_filled(8, 1, true);
-        assert_eq!(grid.get(Pos::new(0, 0)), Some(true));
-        assert_eq!(grid.get(Pos::new(7, 0)), Some(true));
-        assert_eq!(grid.get(Pos::new(8, 0)), None);
-        assert_eq!(grid.get(Pos::new(0, 1)), None);
-    }
-
-    #[test]
-    #[should_panic(expected = "Buffer size does not match grid dimensions")]
+    #[should_panic(expected = "Buffer length must be a multiple of width")]
     fn arr_new_panics() {
-        let _ = ArrayBits::<u8, 1, RowMajor>::new(9, 1);
+        let _ = GridBits::<u8, _, RowMajor>::new(9, 1);
     }
 
     #[test]
     fn impl_slice() {
         let data: [u8; 1] = [0b0000_0001];
-        let grid = SliceBits::with_buffer_row_major(&data, 8, 1).unwrap();
+        let grid = GridBits::<_, _, RowMajor>::from_buffer(data, 8);
 
         assert_eq!(grid.get(Pos::new(0, 0)), Some(true));
         assert_eq!(grid.get(Pos::new(1, 0)), Some(false));
@@ -375,7 +304,7 @@ mod tests {
     #[test]
     fn impl_slice_mut() {
         let mut data: [u8; 1] = [0b0000_0001];
-        let mut grid = SliceMutBits::with_buffer_row_major(&mut data, 8, 1).unwrap();
+        let mut grid = GridBits::<_, _, RowMajor>::from_buffer(&mut data, 8);
 
         assert_eq!(grid.get(Pos::new(0, 0)), Some(true));
         assert_eq!(grid.get(Pos::new(1, 0)), Some(false));
@@ -392,7 +321,7 @@ mod tests {
 
     #[test]
     fn vec_new() {
-        let grid = VecBits::<u8, RowMajor>::new(8, 1);
+        let grid = GridBits::<_, alloc::vec::Vec<u8>, RowMajor>::new(8, 1);
         assert_eq!(grid.get(Pos::new(4, 0)), Some(false));
         assert_eq!(grid.get(Pos::new(7, 0)), Some(false));
         assert_eq!(grid.get(Pos::new(8, 0)), None);
@@ -400,32 +329,16 @@ mod tests {
     }
 
     #[test]
-    fn vec_new_filled() {
-        let grid = VecBits::<u8, RowMajor>::new_filled(8, 1, true);
-        assert_eq!(grid.get(Pos::new(4, 0)), Some(true));
-    }
-
-    #[test]
-    #[cfg(debug_assertions)]
-    #[should_panic(expected = "Buffer size does not match grid dimensions")]
-    fn with_buffer_unchecked_panics() {
-        let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        // width * height = 6
-        // but data.len() = 5
-        let _ = unsafe { VecBits::<_, RowMajor>::with_buffer_unchecked(data, 2, 3) };
-    }
-
-    #[test]
+    #[should_panic(expected = "Buffer length must be a multiple of width")]
     fn out_of_bounds() {
         let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        let grid = VecBits::with_buffer_row_major(data, 8, 2);
-        assert!(grid.is_err());
+        let _ = GridBits::<_, _, RowMajor>::from_buffer(data, 9);
     }
 
     #[test]
     fn into_inner() {
         let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        let grid = VecBits::with_buffer_row_major(data, 8, 1).unwrap();
+        let grid = GridBits::<_, _, RowMajor>::from_buffer(data, 8);
         let (buffer, width, height) = grid.into_inner();
         assert_eq!(width, 8);
         assert_eq!(height, 1);
@@ -436,7 +349,7 @@ mod tests {
     #[test]
     fn iter() {
         let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        let grid = VecBits::with_buffer_row_major(data, 8, 1).unwrap();
+        let grid = GridBits::<_, _, RowMajor>::from_buffer(data, 8);
         let mut iter = grid.iter();
         assert_eq!(iter.next(), Some(true));
         assert_eq!(iter.next(), Some(false));
@@ -450,26 +363,9 @@ mod tests {
     }
 
     #[test]
-    fn with_buffer_row_major_unchecked() {
-        let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        let grid = unsafe { VecBits::<u8, RowMajor>::with_buffer_row_major_unchecked(data, 8, 1) };
-        assert_eq!(grid.get(Pos::new(0, 0)), Some(true));
-        assert_eq!(grid.get(Pos::new(1, 0)), Some(false));
-        assert_eq!(grid.get(Pos::new(8, 0)), None);
-        assert_eq!(grid.get(Pos::new(0, 1)), None);
-    }
-
-    #[test]
-    fn with_buffer_width_exceeded_err() {
-        let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        let grid = VecBits::with_buffer_row_major(data, 8, 2);
-        assert!(grid.is_err());
-    }
-
-    #[test]
     fn as_ref() {
         let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        let grid = VecBits::with_buffer_row_major(data, 8, 1).unwrap();
+        let grid = GridBits::<_, _, RowMajor>::from_buffer(data, 8);
         let slice: &[u8] = grid.as_ref();
         assert_eq!(slice.len(), 1);
         assert_eq!(slice[0], 0b0001_0001);
@@ -478,7 +374,7 @@ mod tests {
     #[test]
     fn as_mut() {
         let data: alloc::vec::Vec<u8> = alloc::vec![0b0001_0001];
-        let mut grid = VecBits::with_buffer_row_major(data, 8, 1).unwrap();
+        let mut grid = GridBits::<_, _, RowMajor>::from_buffer(data, 8);
         let slice: &mut [u8] = grid.as_mut();
         assert_eq!(slice.len(), 1);
         assert_eq!(slice[0], 0b0001_0001);