+++ src/lib.rs
@@ -12,7 +12,7 @@
 //! ## Examples
 //!
 //! ```rust
-//! use grixy::{core::Pos, buf::VecGrid, ops::{GridRead, GridWrite}};
+//! use grixy::{core::Pos, buf::GridBuf, ops::{GridRead, GridWrite}};
 //!
 //! #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 //! enum Tile {
@@ -20,7 +20,7 @@
 //!    Wall,
 //! }
 //!
-//! let mut grid = VecGrid::new_filled_row_major(10, 10, Tile::Empty);
+//! let mut grid = GridBuf::new_filled(10, 10, Tile::Empty);
 //! grid.set(Pos::new(5, 5), Tile::Wall).unwrap();
 //! assert_eq!(grid.get(Pos::new(0, 0)), Some(&Tile::Empty));
 //! assert_eq!(grid.get(Pos::new(5, 5)), Some(&Tile::Wall));